+++ lib/src/components/loading/impaktfull_loading.dart
@@ -3,7 +3,14 @@ import 'package:impaktfull_ui/src/theme/impaktfull_theme_localizer.dart';
 import 'package:lottie/lottie.dart';
 
 class ImpaktfullLoadingIndicator extends StatelessWidget {
+  final String? asset;
+
   const ImpaktfullLoadingIndicator({
+    super.key,
+  }) : asset = null;
+
+  const ImpaktfullLoadingIndicator.custom({
+    required this.asset,
     super.key,
   });
 
@@ -11,7 +18,7 @@ class ImpaktfullLoadingIndicator extends StatelessWidget {
   Widget build(BuildContext context) {
     return ImpaktfullThemeLocalizer(
       builder: (context, theme) => Lottie.asset(
-        theme.assets.lottie.loading,
+        asset ?? theme.assets.lottie.loading,
         width: 48,
         height: 48,
       ),