+++ lib/src/util/extensions/text_style_extensions.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+extension TextStyleExtension on TextStyle {
+  TextStyle withOpacity(double opacity) =>
+      copyWith(color: color?.withOpacity(opacity));
+}